# Make accept use minimal_dispatch