@piscisaureus 2019-05-03
I don't see a significant difference in performance, plus this is a regression in the sense that error information is lost.
So I'm -1 for now.