@ry 2019-05-03
@piscisaureus This takes flatbuffers completely out of the loop for `tests/http_bench.ts` which allows us to focus on the remaining bottlenecks. Without thinking thru the logic of the server is difficult.

The error codes will be dealt with eventually - I think the regression is worth it