@ry 2019-05-03
nevermind. the accept isn't in the hot loop for test/http_bench.ts I will put this off until we have a more established minimal dispatch